fix: 관리자 회원탈퇴 기능 수정
멤버id 로그인 된 상태로 해당 회원 탈퇴 시도 시 alert 띄워주고 탈퇴불가처리
추가로 memberDetail에 신고받은 댓글들 띄우게 기능추가하였으나.. 아직 완성 안 됨

+++ src/main/java/green/with/dodo/SecurityController.java
@@ -1,5 +1,7 @@
 package green.with.dodo;
 
+import java.util.List;
+import java.util.Map;
 import java.io.IOException;
 
 import javax.servlet.ServletException;
@@ -15,9 +17,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import service.ChallengeService;
 import service.MemberService;
+import service.ReplyService;
 import vo.MemberVO;
 import vo.PageVO;
+import vo.ReplyVO;
 
 @Controller
 public class SecurityController {
@@ -25,14 +30,21 @@ public class SecurityController {
 	@Autowired
 	MemberService service;
 	
+	@Autowired
+	ReplyService rService;
+	
+	@Autowired
+	ChallengeService cService;
+	
 	@GetMapping("/adminLogin")
 	public String adminLogin(HttpServletRequest request, String message, Model model) {
-		
+		/*
 		// 기존 member용 아이디 logout
         HttpSession session = request.getSession(false);
 		if (session != null) {
 			session.invalidate();
 		}
+		*/
 		
 		if(message != null) {
 			model.addAttribute("error", "관리자 로그인 에러 : 계정확인");
@@ -96,23 +108,51 @@ public class SecurityController {
 		mv.addObject("selectedMember", vo);
 		mv.setViewName("admin/memberDetail");
 		
+		List<ReplyVO> rvo = rService.reportedList(vo.getNick());
+		
+		
+		mv.addObject("reportedReplies", rvo);
+		
 		return mv;
 	}
 	
     @RequestMapping(value = "/adminMdelete")
     public ModelAndView adminMdelete(HttpServletRequest request, 
-    		ModelAndView mv, MemberVO vo, RedirectAttributes rttr) {
+    		ModelAndView mv, MemberVO vo, RedirectAttributes rttr, PageVO<MemberVO> pvo) {
     	
     	String id = vo.getId();
+    	String loginid = null;
+    	
     	vo.setId(id);
     	
-    	if(service.delete(vo) > 0) {
-    		rttr.addFlashAttribute("message", "회원이 탈퇴처리되었습니다.");
-    	} else {
-    		rttr.addFlashAttribute("message", "탈퇴 오류");
+    	HttpSession session = request.getSession(false);
+
+    	if (session != null && session.getAttribute("loginID")!=null ) {
+    		loginid = (String)session.getAttribute("loginID");
+    		if(!id.equals(loginid)) {
+    			if(service.delete(vo)>0) {
+    				System.out.println("탈퇴성공: " + id);
+    				mv.addObject("message", "회원이 탈퇴처리되었습니다.");
+    				mv = adminMList(mv, pvo);
+    			} else {
+    				System.out.println("service.delete실패");
+    				mv.addObject("탈퇴 오류");
+    			}
+    		} else {
+    			System.out.println("아이디 똑같아서 탈퇴불가");
+				mv.addObject("message", "현재 로그인 중인 계정은 탈퇴처리할 수 없습니다. 회원 계정을 로그아웃하세요.");
+    		    try {
+					mv = memberDetail(request, mv, vo);
+				} catch (ServletException e) {
+					e.printStackTrace();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+    		}
+    		
     	}
     	
-    	mv.setViewName("admin/memberList");
+    	//mv.setViewName("admin/memberList");
     	return mv;
     } //mdelete
